fix remarks pr, improve tests
