+++ templates/home.html
@@ -15,7 +15,7 @@
             <div class="row">
                 <div class="input-field col s12 center-align">
                     <button class="btn waves-effect waves-light light-blue darken-1" type="submit" name="action">
-                        {{.Locale.ButConnect}}
+                        {{.Locale.ButtonSave}}
                         <i class="material-icons right">sync</i>
                     </button>
                 </div>
