+++ translate/translate.en.yml
@@ -1,4 +1,4 @@
-but_connect: Save
+button_save: Save
 tab_settings: Connection settings
 tab_bots: Bots
 table_name: Bot name
