+++ translate/translate.ru.yml
@@ -1,4 +1,4 @@
-but_connect: Сохранить
+button_save: Сохранить
 tab_settings: Настройки CRM
 tab_bots: Боты
 table_name: Имя
